requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
discord.py==2.1.0
frozenlist==1.3.3
h11==0.14.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
osrparse==6.0.2
ossapi==3.0.0
pydantic==1.10.4
python-dotenv==0.21.1
requests==2.28.2
requests-oauthlib==1.3.1
typing-utils==0.1.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
yarl==1.8.2
